refactor: normalize core agent filenames
