--- core/agents/cto.py
@@ -1,1 +1,0 @@
-from core.agents.cto_agent import CTOAgent  # re-export for canonical path
