--- core/agents/ip_analyst.py
@@ -1,1 +1,0 @@
-from core.agents.ip_analyst_agent import IPAnalystAgent  # re-export
