--- core/agents/marketing.py
@@ -1,1 +1,0 @@
-from core.agents.marketing_agent import MarketingAgent  # re-export
