--- core/agents/regulatory.py
@@ -1,1 +1,0 @@
-from core.agents.regulatory_agent import RegulatoryAgent  # re-export
