--- core/agents/research_scientist.py
@@ -1,1 +1,0 @@
-from core.agents.research_scientist_agent import ResearchScientistAgent  # re-export
