+++ dr_rd/agents/synthesizer_agent.py
@@ -1,1 +1,1 @@
-from core.agents.synthesizer import SynthesizerAgent  # noqa: F401
+from core.agents.synthesizer_agent import SynthesizerAgent  # noqa: F401
